[ts2pant-iteration-builder-completeness] Patch 1: move checker stubs to integration
